Fix expansion
A list of expressions is expanded by expanding the head of the list.
If the result is NIL, the tail is expanded as a list. If the result is
the name of a macro, the macro is called with the remaining elements.
Otherwise each of the remaining elements are expanded.